requirements.txt
django==2.2

django-allauth==0.42.0

django-ckeditor==5.9.0

Pillow==7.2.0

django-widget-tweaks==1.4.8

django-crispy-forms==1.9.2

django-cleanup==5.0.0

